Rename local_wasm to wasm
